Fix: Tailwind error about nested CSS when React rendered
I knew that this error is about devDependencies, but not.
It would occur when Tailwind CSS and postcss-nesting used together.
It is thought that using nesting functions more than one is rejected.
The solution is either
removing 'tailwind.config.js' to use postcss-nesting
or uninstall postcss-nesting to use only Tailwind.
I selected to uninstall postcss-nesting
because Tailwind CSS already have same function.

Resolves: #22